Passge des double en int pour StatusModel

+++ EZSave/Models/StatusModel.cs
@@ -4,9 +4,9 @@
     {
         public string Name { get; set; } = "";
         public string State { get; set; } = "";
-        public double Progress { get; set; }
-        public double Size { get; set; }
-        public double TotalTransferTime { get; set; }
+        public int Progress { get; set; }
+        public int Size { get; set; }
+        public int TotalTransferTime { get; set; }
         public int TotalFilesToCopy { get; set; }
     }
 }
